refactor: 인기 검색어 영역 컴포넌트 분리
