+++ app/(main)/search/_modules/index.ts
@@ -1,4 +1,6 @@
 export * from './SearchInput';
+export * from './SearchPageClient';
+export * from './SearchResults';
+export * from './TrendingSearches';
 export * from './TrendingSearchItem';
 export * from './types';
-export * from './SearchResults';
